delete user routes
